Introduce markdown editor and renderer.
And apply it to Issue Tracker.

Usage
-----

Call @views.html.markdown in your scala template, to enable the markdown
support.

@views.html.markdown accepts a map as a parameter. The map should have
keys as selectors for html elements and values as the actions, "edit" or
"render", will be done for the each element.

Here is an example:

    @import scala.collection.mutable.Map
    @views.html.markdown(Map("#body" -> "edit"))

See app/views/issue/editIssue.scala.html for a real example.