shortcut: ctrl + enter to submit a form.
Add this shortcut key for the Issue Tracker.

How to enable the submitting shortcut key in the any other page
---------------------------------------------------------------

Here is an example. The below makes all forms in the html page to
support ctrl+enter as a submitting shortcut key.

    <script>
      nforge.require('shortcut.submit');
    </script>

See app/views/issue/editIssue.scala.html for a real example.