issue: bugfix for downloading issues as excel.
An internel server error has been occured while trying to download
issues as a excel file because of the file does not exist. The reason of
the error is lack of the directory in which the excel file is saved, so
the error may be fixed easily by just creating the directory.

But I have found that saving excel file in a directory is not safe at
all because the directory is public so cannot be access-controlled. To
solve this problem, I have changed the form of excel data generated by
Issue model, from a regular file to byte array, to make them be in the
memory, not a directory.