issue: Use more proper status code 204 than 200.
When user request to attach a label to an issue and the label is already
attached, the server response with empty body because there is no label
added newly. In this case, 204 No Content is better status code than 200
because it makes sure the empty body is not a mistake and inteded.

In Addition, This situation is not an error because it means, "Attaching
label is successfully, but there was no need to add label newly to the
issue because the label had been already attached to the issue."