PullRequest: Fix an error on outdated comment.
While rendering an outdated comment, PullRequestComment updated the
editList of the diff the comment attached to render interest portion
only from the entire diff. But it was an inappropriate behavior because
it occurs errors between the editList and rawtext.

After applying this patch, while rending an outdated comment,
PullRequestComment updates fileDiff to make fileDiff.getHunks to return
the interest hunk only. It is slower than before because of unnecssary
finding of hunks, but works without error.