Review: Fix NPE when see a thread on outdated changes
Reason: NPE occurs in PullRequestCommit.getStateByCommitId because there
is no PullRequestCommit instance matches the given commitId if the id is
for outdated merge commit of the pull request.

Solution: Consider a thread on a commit which belongs to a pullrequest
but not to the pullrequest's pullRequestCommits as a thread on outdated
changes, not on a specific commit, of the pullrequest.

Note: Many lines of code added into CodeCommentThread are imported from
PullRequestComment which had been removed at 4132012.