docs/logging: Update for access log format
Add description about the time taken to server the request.

It may be a little bit confused because the time taken to build and
write response body is not taken account into the time taken to serve
the request when chunked encoding is used. So this description is
required to make it clear.