code: Improve the error handling for oversized diff
Now, If one or both of the files are too big to show or get the
difference from them, Yobi show an error message different from
one for the case if the difference is oversized.

This also fixes an error have occured when render small diff of an
oversized file because FileDiff.getError returns A_SIZE_EXCEEDED or
B_SIZE_EXCEEDED if the file is larger than the FileDiff.LINE_LIMIT. If
the diff is not larger than the limit, it should be able to rendered
even if the file is larger than the limit but small enough not to raise
org.eclipse.jgit.errors.LargeObjectException.

And add tests for some exceptional cases.