Review: Fix sql error of 72.sql caused by 4ea69ee
Reason: The tables for Review feature, review_comment and
comment_thread, MUST NOT be deleted by 72.sql because they are NOT
added by the script. This causes sql error while running the ups part of
72.sql because it assumes the tables exist.

Solution: Delete every rows of the tables instead drop them if need to
avoid the duplication of review comments.