Review: Fix missing attachment on review comment
Problem: Files are not attached to review comments except the first one
of its thread.

Reason: _setReviewFormFields() of yobi.CodeCommentBox.js removes
'temporaryUploadFiles' field in the comment form when show the form and
_addUploadFileIdToListAndForm() of yobi.Attachments.js does not add
any files into the list of files to be attached if the field does not
exist.

Solution: Make _setReviewFormFields() not to remove the field.