issue: Fix NPE while fixing last issue number
Every project maintains "last issue number". When someone submit a
new issue, Yobi set the issue's number to the last issue number + 1.
7be78b0 introduces a guard code to prevent that Yobi set incorrect
number. The guard fixes the issue number if the last issue number is
smaller than number of issues of the project.

But it did not work at all because NPE occurs during the fixing. The
reason is trying to get a 'number' property from a latest issue and
assign it to primitive long type variable but the property's value is
null.

Why the value is null and how to fix it? Here are the reasons and
solutions:

1. The issue is already in persistent context. If you use @Transaction
   annotation and try to get a bean using Ebean in the transaction, it
   returns a bean with null properties if the bean was already in
   persistent context(it means you already got the bean somewhere).

   Solution: Refresh the issue before getting a property.

2. The fixing feature assumes number property of issue is not null. But
   it is not true. In unitteset, an issue with null number is common.

   Solution: Set 0 to the last issue number if latest issue's number is
   null.