git: Fix an error while clonning a forked project
WARNING: THIS MAY CAUSE A SERIOUS PROBLEM IN SOME GIT REPOSITORIES
MANAGED BY YOBI. YOU NEED TO READ THE LAST PARAGRAPH BEFORE APPLYING
THIS.

Problem: MissingObjectException may occur while clonning a forked
project if some objects of the origin project are in its
merge-repository but not in the origin project so they were missed while
clonning.

Reason: Using alternate object database helps lookup some objects but
may not store some pushed objects, which are in the alternate object
database, in the repository.

Solution: Do not use alternate object database while pushing.

Warning: This may break pushing to some git repositories which have a
merge-repository because of missing object, with an error as follows:

    To http://yobi/you/repo1
    ! [remote rejected] master -> master (object 97c4c34b825debd0fe76bf5d850d0891d4a93bba missing)
    error: failed to push some refs to 'http://yobi/you/repo1'

Fortunately, the repositories can be recovered by copying git objects
from the object directory of the merge-repository. For an example:

    cd yobi
    cp repos/git-merging/you/repo1.git/.git/objects/* \
       repos/git/you/repo1.git/objects -R