CodeBrowser: Add notice for large file
When a user tries to show a large file in CodeBrowser, show a notice to
tell that we can't show the file because of the size and a button to
show the file in raw. The button means "You can see the file in raw
instead".

If the user is a Site administrator, tell also that you can loose the
size limit.