Home: performance tuning
The event view loads all notificationEvents related with a user and slice it. Also, this operation is called twice in it.

First of all, I've fixed to load notificationEvents that only to show in this view. Last but not least, I've refactored the view to use @defining to load the events only once.