Group admins or members are listed on issue's Assignee list.
Cause:
There is VOC.
When writing an issue on a protected project, we need to assign the issue to someone who is group member. But assignee list includes only project members.

Solution:
This already has been solved. 1ba9698
But there is a bug that Group admins are not listed on private project which has a group.
So group admins are added to assignee list on private project's issue.

Private-issue: 1297