Changes width and align of `Commit Date` column on `Code` menu.
Cause:
`Commit Date` string, a column title, can not be shown as one line on a div following `span1` class because of long.

Solution:
A class which is stylesheet for the column title is changed to `span2` class.
Also `text-align` of all the column's data is set to `right`.

Private-issue: 1518