PullRequest: fixed a bug in the code-reivew tab
When a new pull-request is created and a user click code-reivew tab quickly before creating mergedCommitIdFrom and mergedCommitIdTo, a server error occures because it cannot generate diff without those variables that are going to be created after attemptMerge finished.

I've added 'isDiffable' method that can be used to know if it can generate diffs or not. And I used it before showing diffs in the code-review tab.