PullRequest: Fix painful waiting for merging
Reason: When Yobi merges a pullrequest, it also gets authors related to
the changes by the pullrequest to send them notification. But it is very
very slow because jgit-blame is about 8x slower than git. It takes
hundreds of milliseconds for each changed file.

Solution:
* Get related authors just in time (just before sending notification).
* Give up getting related authors from changes if the changes include
  too many files. The limit is defined as
  GitRepository.BLAME_FILE_LIMIT.

And remove PullRequest.relatedAuthors field because this patch removes
the need.