Attachment: Allow HTTP cache
Set Cache-Control header to "private, no-store, max-age=3600".

I chose the directives based on following reasons:

* "private" avoids that attachments whose access is restricted are
  shared publicly.
* "no-store" reduces possibility that sensitive attachments are leaked
  to a person who has no permission. e.g. A member of private project
  might sell his laptop. The member believed sensitive files are deleted
  by self completely but the web browser's cache may still have
  sensitive attachments.
* See warning below for "max-age=3600".

Warning: This patch introduces a side-effect that a user still access
attachments, from the user's private cache, which the user "had" the
permission but not now. "max-age=3600" directive helps to relieve the
problem by telling the cache to expire the attachment which is older
than 3600 seconds.