issue: Fix an OptimisticLockException while saving
Problem: f3e0112 has a bug that throws an OptimisticLockException while
saving new issue in a project whose lastIssueNumber is incorrect.

Reason: increaseLastPostingNumber() tried to update a project which is
updated already fixLastIssueNumber().

Solution: Do not call fixLastIssueNumber() in
increaseLastPostingNumber(). Instead make project.getLastIssueNumber()
to correct the number if necessary.

Apply this patch to also postings.