Shows a assignee on assignee dropbox correctly.
Cause:

If users click `Assign to me` or `Assign to author` twice, a assignee is not shown as a user name. It will be displayed as `Assign to me` or `Assign to author`.

The assignee on assignee dropbox will be changed to current user after a user clicks `Assign to me`. Becuase the current assignee is not same to current user. But the assignee will not be changed to current user when a user click `Assign to me` once again. Because `Assign to me` option tag's value is same as current user id, changed event is not occur. So the assignee is displayed as `Assign to me`.

Solution:

`select2-selecting` event is added. This is fired when a choice is being selected in the dropdown, but before any modification has been made to the selection.

When selected items are `Assign to me` or `Assign to author`, the event is fired and the event calls the change event to the dropdown forcibly. So the assignee is changed to the proper user.

Private-issue: 1313