mail: Add 'rel=noreferrer' to external links
Add 'rel=noreferrer' to every external link only if
application.noreferrer is true. It requests clients not to add
Referer header in the request to the link.

We assume a link is a external if the hostname is not empty and does not
equal to the system's hostname from Config.getHostname().

Note: It works only for clients which support HTML5.