`Assign to me` and `Created by me` are not shown when current user is anonymous.
Cause:
Although current user is anonymous, `Assign to me` and `Create by me` are shown on `Issue` menu. But it is nonsense. Anonymous can not write any issues and not be a assignee.

Solutions:
When current user is anonymous, `Assign to me` and `Create by me` will not be shown.
For these modifications, actually, `!UserApp.currentUser().isSiteManager` is replaced to `!UserApp.currentUser().isAnonymous`. The previous codes are bug and new codes are to solve this issue.

Private-issue: 1676