code: Fix incorrect listing of files in subdirectory, again
This bug is a jgit bug and I worked around it at c4f88e9 by creating
a new iterator for commits when incorrect behavior of jgit is detected.
But I made a mistake that stops iterating if the current commit is same
with the previous one. I shouldn't have done that because the new
iterator may produce a same commit as the previous one.