fixes `Unwatch link` on notification emails which does not work.
Cause:

Unwatching issues does not work when clicking `Unwatch link` on notification emails about creating issue comments. `Unwatch link` is made with a resource that occurs with a notification email. But `Unwatch link` should be made with a resource, issue,  when creating issue comments.

Solutions:

When creating issue comments, `Unwatch link` will be made with the issue which contains the issue comments.

Private-issue: 1787