PullRequest: Do not make merging repository
Merge a pullrequest without a non-bare repository. It is possible
because JGit allows to merge branches in memory. The idea comes from
a commit of Gitbucket [1].

Instead of the merging repository, Yobi uses internally some refs which
match with these three patterns:

* "refs/yobi/pull/:id/head" to indicate the fetched source branch
* "refs/yobi/pull/:id/merged" to indicate the merged commit
* "refs/yobi/pull-check/:project-owner/:project-name/:branch" to
  indicate the fetched source branch to be used to check conflict before
  sending a pullrequest.

To reserve them, the new pre-push hook "RejectPushToReservedRefs"
rejects pushes to "refs/yobi" and any branch which starts with
"refs/yobi/".

Benefits: This saves disk space, makes merging faster and removes the
need to maintain merging repository.

Backward Compatibility: This patch does not delete existing merging
repositories and the codes to access them because the repositories are
still required for pullrequests already sent.

Side Effect: Unfortunately the list of conflicted files are no more provided
because the new merged which stops immediately when a conflict occurs
can't gather them.

[1]: https://github.com/takezoe/gitbucket/commit/af8a1234ede105c0ea011a6e21ad881810cd4cb6