Group: Fixed to show group's projects
Some projects were not shown even to users who has right to see them.

I'm not sure what was the reason, the project's owner value and the group's name value
has different case for the first letter.

This commit contains a sql that finds all those projects and set the right organization id value to it.