shows a alert view when site admin removes a member who is only a project's manager.
Cause:
When the site admin removes a sole project's manager on a project, any alerts or messages did not occur. The delete method is called by AJAX but Yobi server returns redirect response. It can not show alert view related to `flash method`.

Solution:
Proper HTTP response codes was added and a error callback function was inserted to javascript codes. The callback function will compare responses and if the reponses are 401 or 403, current page is reloaded to show a alert view.

Private-issue: 1659