Shows non-member assignees on assignee dropbox.
Cause:
If a member is removed from a project after the member is assigned to an issue, the member is not listed on assignee dropbox. Because assignee list contains only project and group members.

Solution:
Although some assignee are not member of projects and groups, those will be shown on assignee dropbox because of assignees.

Private-issue: 1667