fix not showing comment layer in closed thread
If the fragment identifier component of a URI is comment layer dom element id, the comment thread layer is not showing.

The onhashchange function invoke _scrollToAndhighlight function.
The _scrollToAndhighlight function was implemented to show the comment thread layer dom element received as argument.
But if this argument is comment layer dom element, this function didn't show the comment thread layer included comment dom element.

So if this argument is not the comment thread dom element,
I've changed the _scrollToAndhighlight function to find the comment thread dom element included comment layer dom element id.