Introduce Mailbox Service
Mailbox Service is a service which fetches emails from the IMAP server
configured by imap.* and posts the emails.

If the email address to yobi is "yobi@mail.com":

* If a user want to posts an issue in project "my/proj", just send an
  email to "yobi+my/project@mail.com". The email will be posted as an issue.
* Users can post a comment by replying to notification emails. It works
  only if the notification email can accept replies; The footer of the
  notification email describes the ability.
* If a user discuss using email with many people and you want that Yobi
  records the discussion as an issue and comments in project "my/proj",
  just add an address "yobi+my/proj@mail.com" in CC list. Note: But Yobi
  will records only the emails from Yobi users.
* If a user don't know how to send emails he or she can send an email to
  "yobi+help@mail.com". Then yobi replies with help messages.

Mailbox Service fetches emails in two ways: polling and listening.

* Listening - Yobi listens updates from the IMAP server and
  process new emails immediately. Listening works only if the IMAP
  server supports IDLE command.
* Polling - Yobi fetches new emails on the interval configured by
  `application.mailbox.polling.interval`.

Yobi tries listening first and only if it fails tries polling.

If you want to know how to add the configuration for the IMAP server,
see the Mailbox Service section in conf/application.conf.default file.