attachments: Optimize getting list of attachments
Problem: Yobi sent a request to get a list of attachments for each
comments. It issued huge amount of requests when the user read an issue
with many comments.

Solution: Embed the list of attachments as a value of data-attachments
attribute and render the list by using javascript.

I expect the total requests will reduce by 41%.