ProjectApp: Fix the incorrect status code for label
When a user tries to attach a label to a project, ProjectApp.attachLabel
responses 201 if the label is created newly and attached and 204 if the
label is already attached.

But it shouldn't have work correctly because it was not possible to know
whether the label has been already attached if the label is created just
now. I have no idea why this has worked correctly until I upgrade Play
to 2.2.6.

This also fixes the condition to warn that a label is not attached
abnormaly.