fixes `Assign to me` and `Create by me` on issue view.
* Issue

When one user is not a project member and manager and is a group member or manager, `Assign to me` and `Create by me` do not work. Because current user is not listed on assignee's and author's combo box.

* Solution

Current user will be listed on assignee's and author's combo box.

Private-issue: 1849