fixes a bug that current user avatar is not shown on project list.
* ISSUE

Current user avatar is not displayed on the right of `Project List` intentionally. Becuase it is shown on the left side. But now it is not shown and project avatar is shown in there. So the user avatar should be appeared.

* SOLUTION

Current user will be shown on `Project List`.

Private-issue: 956