changes queries for performance.
* Issue

When issue page is loaded, lists of assigness and authors are slow.
A statement to add current user to a query result is problem.

* Solution

The statement is removed and current user is added to the result manually.
It is 16.4 times faster than v0.8.0-rc1. (v0.8.0-rc1 avg:5008ms, current commit avg:304.8ms)

Private-issue: 1849