Don't use version.conf
Problem: We should not store Yobi's version information in
conf/version.conf because user may not update the file when upgrade
Yobi. Since Play 2.3, conf directory for product is copied while
packaging so the directory is not updated even if Yobi is upgraded.

Solution: Store the version information in build.sbt. We used
version.conf to know the application version, required for YobiUpdate
feature, while application is running. But it is not necessary because
we can get the version by using sbt-buildinfo.