Notification: Fix message-id duplication for pushing commits
If a notification is caused by an event about "creating something", we
give id of the resource as message-id of the notification email so that
Yobi can create comments on the resource, which is determined by the
message-id in In-Reply-To header, from the replies.

But if Yobi sends multiple "creating" notifications on a same resource,
it will send multiple emails with same message-id and it may cause
problems in MUA, e.g. trash them.

Notification for pushing commits is that. It was considered as "creating"
event but the resource is "project" even if the event does not create
a project. Naturally, pushing commits to a same project is very common
event so it has been caused the problem very frequently.

Solution: Do not consider "NEW_COMMIT" as "creating" event.