ProjectTransfer: fixed bug
Problem: AccessControl is allowed all requests to site admin or member of the project in isProjectResourceAllowed method,
so a user of a project that should be transfered to other user or group, can accept the transfer request.

Solution: Changed isProjectResourceAllowed method to check the current user can accept the project-tansfer request
prior to check the current use is admin of the group or member of the project.