fixes logics splitting and sending notification emails with recipient limit value.
* Issue

When `application.notification.bymail.hideAddress` is true, a default recipient is added to `to header`.
But previous codes does not consider it so notification mails, `application.notification.bymail.recipientLimit` + 1, are sent.

* Solution

The changed codes decide amount of notification mails to be sent with `application.notification.bymail.recipientLimit` and recipients in `to header`.

Private-issue: 2206