NotificationMail: Merge mails if possible
When a user leaves a comment on an issue or review thread and close or
reopen it immediately, two events are created; one for comment and
another one for closing or reopening.

In such a case, Yobi have sent two emails to notify the two events. But
this patch makes Yobi sends only one email which has the message
generated by merging the two message of the two events.

For example, until now you have received two emails in such a case as
follows:

    From: Yi EungJun <eungjun.yi@navercorp.com>
    To: me <me@mail.com>
    Subject: [yobi] awesome feature request (#123)

    The feature is already implmented!

    ...

    From: Yi EungJun <eungjun.yi@navercorp.com>
    To: me <me@mail.com>
    Subject: [yobi] awesome feature request (#123)

    Issue has been closed

    ...

But now you receive only one email:

    From: Yi EungJun <eungjun.yi@navercorp.com>
    To: me <me@mail.com>
    Subject: [yobi] awesome feature request (#123)

    The feature is already implmented!

    ---

    Issue has been closed

    ...

Solution: Merge emails by NotificationMail.mergeMails() before they are
sent. The merged emails are instances of MergedNotificationMail whose
interface is same with existing NotificationMail so we didn't need to
modify the existing logic to send notification emails.

Note: Unfortunately, Two emails whose recipients differ from each other
are not merged. It means that if someone watches a project and changes
his or her notification setting not to receive notifications for new
comment then this feature does not work on the project because the user
exists in recipients of close/reopen events but not new comment events.