NotificationMail: Merge mails even if recipients differ
If two mails can be merged but their recipients differ from each other,
split them into three notifications.

For example, if there are two notifications as follows:

    * a notification of new comment to foo and bar
    * a notification of issue closing to bar and baz

Send three notifications as follows:

    * a notification of new comment to foo
    * a notification of issue closing to baz
    * a merged notification of new comment and issue closing to bar

Remove INotificationMail and MergedNotificationMail. They were needed
not to modify the logic to send email but the modification is inevitable
to apply this patch so we do not need them anymore.

Side effect: "application.notification.bymail.hideAddress = false"
becomes less helpful. Before applying this patch, when a user receives
a notification, the user could guess who does not receive it from "To"
header. But now it is quite difficult because the notification could be
merged into another notification.