title-header: Add header filter at title of posting
Users commonly use [ ] in their issue/board/pull-request
title to define category. We recommended using labels, but -anyway- if
users prefer to use it, I thought that it would be better
to help using it conveniently.

Now, header words of title will be appeared with link color and
work like a filter link. When a user click the header word,
issue/board/pull-request list of same header will be displayed.