NotificationMail: Fix ConcurrentModificationException
A ConcurrentModificationException occured because an iterator to the
list of events, to which an event was added after the creation of the
iterator, was accessed. This bug was caused at dfcf344.

Side effect: The ordering of events is changed. For example, if there
are three events:

* Event 1: User A commented on the issue
* Event 2: User B commented on the issue
* Event 3: User A closed the issue

and Event 1+3 is the merged event of 1 and 3, then they are sent in the
order of 1+3 and 2 instead of the original order, 2 and 1+3.