Milestone: enhance form validation
Change the regular expression to prevent wrong date string like 2015-05-2222.
Even if, the changed regular expression can prevent it, but it is still possible to send wrong date sting like 2015-05-99.
So, I've added server-side validation for the case of someone send form data without browser.
Last but not least, in case there are form errors checked by server-side, show the form data that the user sent.