Fix error pages that don't work for anonymous
Error pages wasn't shown correctly for anonymous. The error pages used
a template in app/views/siteLayout.scala.html which ignored the given
content and rendered always the intro page if the current user is
anonymous.

Decide whether to show the intro page or not in index.scala.html not
siteLayout.scala.html.