Markdown: Set timeout of rendering to 5 seconds
Return the source wrapped by `<pre>` HTML tag, instead of rendered
result, if timeout occurs.

Warning: Thread.stop(), which is danger and deprecated but used in this
commit, may corrupt the script engine. It may be better to add some code
to rebuild the script engine after stopping a thread if you meet some
problem in future.