perf: Avoid unnecessary permission check
Problem: When getting watchers only to know whether a specific user is watching
the resource, checking every user, who watches the resource, has the proper
permission to read the resource is not necessary.

Solution: Add a boolean parameter to getWatcher() methods to tell whether
permission check is necessary.

This makes it two times faster, to get an issue with 19 comments, in a project
watched by 344 users.