perf/user: Make hash() and equals() faster
Do not consider user.loginId. It is not necessary because user.id is enough to
identify the user.

Getting user.loginId may be slow because it may require an additional query to
the database.

This makes it 1.31 times faster, to get an issue with 19 comments, in a project
watched by 344 users.