perf/issue: Improve performance using @Transactional
Put @Transactional annotation to IssueApp#issue() and IssueApp#issues(). It
enables the cache by ORM and improves the performance 2-3 times.

This makes it 6.28 times faster, to get an issue with 19 comments, in a project
watched by 344 users. I am pretty sure that the annotation on IssueApp#issues()
also improves the performance.

Note: We already tried putting @Transactional to the all controllers at
e9d63df, but rollbacked it at 8c25ca9 because of some bugs that some getters of
entities occasionally return null. I guess 13f8014 prevents the bug on IssueApp
so this commit seem to be okay.