attachment: Use SHA-256 instead of SHA-1
Overwriting an existing attachment file maliciously is possible since
Yona uses SHA-1 algorithm, which is known to be shattered[1], to digest
the contents to generate the names of the attachment files.

This fix ensures backward compatibility. Only new attachments have
filenames generated by SHA-256 algorithm. Fortunately, the length of
'name' column of 'attachment' table is 255 which is enough to store
SHA-256 which requires 64.

[1]: https://shattered.it/