notification: Fix getting user display name bug
There is a very rarely but intentionally maden code.

When to get member's display name according to login user's language,
model calls controller, technically Http Context.

So, most of cases it has no problem. But under the some circumstance such as
notification events which is called by akka scheduler, it occurs problem.

This commit fix it.