title-prefix: Support auto-completion title header
Title header and lable auto completion is triggered by '[' character
in the issue/posting title input box.

TL:DR
---
Title prefix header is loved feature of Yona.
It is very conveneint for expressing category of the posting.

But somtimes user mistyped or made duplication of title header which is
nearly the same but not exactly same. And, more Yona users adopt title
header, more they omit selecting issue labels.

It's a bad thing for all of us.

Now, Yona supports auto-completion of title header prefix and issue labels.
This is a trial to recude this kind of problem.

I hope it will be going well.