checklist: Fix authorization related bug
Previously @IsAllowed annotation was used.
But I found there is a bug.  Becuase most of annotation based
authorization is depends on PathParser. But PathParser
is rather unstable because of its url parsing method.

So, for the reason, a user who has authorization can not change state of checklist,
except issue author.

This patch fix that problem.